test (first) slots available + limit appointment by numberOfSlots + command update firstSlotAvailable
In this commit, I test slotsAvailable and firstSlotAvailable update when there are appointment events, opening hours events and exceptional closure events.
There is a new command and service for update firstSlotAvailable for all repairer with firstSlotAvailable in past.
I add a new voter for Appointment POST, to check if slots are available (to avoid to had 50 appointment if a repairer accept only 4 appointment on a opening hour)
I have refactor some tests who needs it.